Fixed a few details.
