+++ src/collective/taxonomy/interfaces.py
@@ -1,8 +1,13 @@
 # -*- coding: utf-8 -*-
 
+from plone.namedfile.field import NamedBlobFile
+
 from zope.interface import Interface
 from zope.i18n.interfaces import ITranslationDomain
+from zope import schema
 from zope.schema.interfaces import IVocabularyFactory
+
+from .i18n import MessageFactory as _
 
 
 class IBrowserLayer(Interface):
@@ -20,3 +25,64 @@ class ITaxonomy(ITranslationDomain, IVocabularyFactory):
 
     def __setitem__(identifier, term):
         """ """
+
+# Control panel stuff
+
+
+class ITaxonomySettings(Interface):
+    """ Schema for controlpanel settings """
+
+    taxonomies = schema.List(
+        title=_(u"Taxonomies"),
+        value_type=schema.Choice(
+            description=_(
+                u"help_taxonomies",
+                default=u"Select the taxnomies you desire to modify"
+            ),
+            required=False,
+            vocabulary='collective.taxonomy.taxonomies',
+        ),
+        default=[],
+    )
+
+
+class ITaxonomyForm(Interface):
+    # Regular fields
+
+    field_title = schema.TextLine(
+        title=_(u"Field title"),
+        required=True
+    )
+
+    field_description = schema.TextLine(
+        title=_(u"Field description"),
+        required=False
+    )
+
+    import_file = NamedBlobFile(
+        title=_(u"Upload VDEX xml file"),
+        description=_(u" "),
+        required=False
+    )
+
+    is_required = schema.Bool(
+        title=_(u"Is required?"),
+        required=True
+    )
+
+    multi_select = schema.Bool(
+        title=_(u"Multi-select field"),
+        required=True
+    )
+
+    write_permission = schema.Choice(
+        title=_(u"Write permission"),
+        required=False,
+        vocabulary='collective.taxonomy.permissions'
+    )
+
+    # Taxonomy hidden field
+    taxonomy = schema.TextLine(
+        title=_(u"Taxonomy"),
+        required=False
+    )
